move permission functionality from HasRoles Trait to HasPermissions Trait
