+++ src/Traits/HasPermissions.php
@@ -4,12 +4,14 @@ declare(strict_types=1);
 namespace Maklad\Permission\Traits;
 
 use Illuminate\Support\Collection;
+use Jenssegers\Mongodb\Eloquent\Builder;
 use Jenssegers\Mongodb\Eloquent\Model;
 use Jenssegers\Mongodb\Relations\BelongsToMany;
 use Maklad\Permission\Contracts\PermissionInterface as Permission;
 use Maklad\Permission\Exceptions\GuardDoesNotMatch;
 use Maklad\Permission\Guard;
 use Maklad\Permission\Helpers;
+use Maklad\Permission\Models\Role;
 use Maklad\Permission\PermissionRegistrar;
 
 /**
@@ -18,6 +20,17 @@ use Maklad\Permission\PermissionRegistrar;
  */
 trait HasPermissions
 {
+    public static function bootHasPermissions()
+    {
+        static::deleting(function (Model $model) {
+            if (isset($model->forceDeleting) && !$model->forceDeleting) {
+                return;
+            }
+
+            $model->permissions()->sync([]);
+        });
+    }
+
     /**
      * A role may be given various permissions.
      * @return BelongsToMany
@@ -184,5 +197,141 @@ trait HasPermissions
         });
 
         return $permissions;
+    }
+
+    /**
+     * Return a collection of permission names associated with this user.
+     *
+     * @return Collection
+     */
+    public function getPermissionNames(): Collection
+    {
+        return $this->getAllPermissions()->pluck('name');
+    }
+
+    /**
+     * Return all the permissions the model has via roles.
+     */
+    public function getPermissionsViaRoles(): Collection
+    {
+        return $this->load('roles', 'roles.permissions')
+            ->roles->flatMap(function (Role $role) {
+                return $role->permissions;
+            })->sort()->values();
+    }
+
+    /**
+     * Return all the permissions the model has, both directly and via roles.
+     */
+    public function getAllPermissions(): Collection
+    {
+        return $this->permissions
+            ->merge($this->getPermissionsViaRoles())
+            ->sort()
+            ->values();
+    }
+
+    /**
+     * Determine if the model may perform the given permission.
+     *
+     * @param string|Permission $permission
+     * @param string|null $guardName
+     *
+     * @return bool
+     * @throws \ReflectionException
+     */
+    public function hasPermissionTo($permission, $guardName = null): bool
+    {
+        if (\is_string($permission)) {
+            $permission = \app(Permission::class)->findByName(
+                $permission,
+                $guardName ?? $this->getDefaultGuardName()
+            );
+        }
+
+        return $this->hasDirectPermission($permission) || $this->hasPermissionViaRole($permission);
+    }
+
+    /**
+     * Determine if the model has any of the given permissions.
+     *
+     * @param array ...$permissions
+     *
+     * @return bool
+     * @throws \ReflectionException
+     */
+    public function hasAnyPermission(...$permissions): bool
+    {
+        if (\is_array($permissions[0])) {
+            $permissions = $permissions[0];
+        }
+
+        foreach ($permissions as $permission) {
+            if ($this->hasPermissionTo($permission)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * Determine if the model has, via roles, the given permission.
+     *
+     * @param Permission $permission
+     *
+     * @return bool
+     */
+    protected function hasPermissionViaRole(Permission $permission): bool
+    {
+        return $this->hasRole($permission->roles);
+    }
+
+    /**
+     * Determine if the model has the given permission.
+     *
+     * @param string|Permission $permission
+     *
+     * @return bool
+     * @throws \ReflectionException
+     */
+    public function hasDirectPermission($permission): bool
+    {
+        if (\is_string($permission)) {
+            $permission = \app(Permission::class)->findByName($permission, $this->getDefaultGuardName());
+        }
+
+        return $this->permissions->contains('id', $permission->id);
+    }
+
+    /**
+     * Return all permissions the directory coupled to the model.
+     */
+    public function getDirectPermissions(): Collection
+    {
+        return $this->permissions;
+    }
+
+    /**
+     * Scope the model query to certain permissions only.
+     *
+     * @param Builder $query
+     * @param string|array|Permission|Collection $permissions
+     *
+     * @return Builder
+     */
+    public function scopePermission(Builder $query, $permissions): Builder
+    {
+        $permissions = $this->convertToPermissionModels($permissions);
+
+        $roles = \collect([]);
+
+        foreach ($permissions as $permission) {
+            $roles = $roles->merge($permission->roles);
+        }
+        $roles = $roles->unique();
+
+        return $query->orWhereIn('permission_ids', $permissions->pluck('_id'))
+            ->orWhereIn('role_ids', $roles->pluck('_id'));
     }
 }
