+++ tests/HasRolesTest.php
@@ -316,159 +316,6 @@ class HasRolesTest extends TestCase
     }
 
     /** @test */
-    public function it_can_determine_that_the_user_does_not_have_a_permission()
-    {
-        $this->assertFalse($this->testUser->hasPermissionTo('edit-articles'));
-    }
-
-    /** @test */
-    public function it_throws_an_exception_when_the_permission_does_not_exist()
-    {
-        $can_logs = [true, false];
-
-        foreach ($can_logs as $can_log) {
-            config('permission.log_registration_exception', $can_log);
-
-            try {
-                $this->expectException(PermissionDoesNotExist::class);
-
-                $this->testUser->hasPermissionTo('does-not-exist');
-            } finally {
-                $message = $this->helpers->getPermissionDoesNotExistMessage('does-not-exist', 'web');
-                $this->assertLogMessage($message, Logger::ALERT);
-            }
-        }
-    }
-
-    /** @test */
-    public function it_throws_an_exception_when_the_permission_does_not_exist_for_this_guard()
-    {
-        $can_logs = [true, false];
-
-        foreach ($can_logs as $can_log) {
-            config('permission.log_registration_exception', $can_log);
-
-            try {
-                $this->expectException(PermissionDoesNotExist::class);
-
-                $this->testUser->hasPermissionTo('admin-permission');
-            } finally {
-                $message = $this->helpers->getPermissionDoesNotExistMessage('admin-permission', 'web');
-                $this->assertLogMessage($message, Logger::ALERT);
-            }
-        }
-    }
-
-    /** @test */
-    public function it_can_work_with_a_user_that_does_not_have_any_permissions_at_all()
-    {
-        // TODO try to use object without saving it to database
-        $user = User::create(['email' => 'new@user.com']);
-
-        $this->assertFalse($user->hasPermissionTo('edit-articles'));
-    }
-
-    /** @test */
-    public function it_can_determine_that_the_user_has_any_of_the_permissions_directly()
-    {
-        $this->assertFalse($this->testUser->hasAnyPermission('edit-articles'));
-
-        $this->testUser->givePermissionTo('edit-articles');
-
-        $this->refreshTestUser();
-
-        $this->assertTrue($this->testUser->hasAnyPermission('edit-news', 'edit-articles'));
-
-        $this->testUser->givePermissionTo('edit-news');
-
-        $this->refreshTestUser();
-
-        $this->testUser->revokePermissionTo($this->testUserPermission);
-
-        $this->assertTrue($this->testUser->hasAnyPermission('edit-articles', 'edit-news'));
-    }
-
-    /** @test */
-    public function it_can_determine_that_the_user_has_any_of_the_permissions_directly_using_an_array()
-    {
-        $this->assertFalse($this->testUser->hasAnyPermission(['edit-articles']));
-
-        $this->testUser->givePermissionTo('edit-articles');
-
-        $this->refreshTestUser();
-
-        $this->assertTrue($this->testUser->hasAnyPermission(['edit-news', 'edit-articles']));
-
-        $this->testUser->givePermissionTo('edit-news');
-
-        $this->refreshTestUser();
-
-        $this->testUser->revokePermissionTo($this->testUserPermission);
-
-        $this->assertTrue($this->testUser->hasAnyPermission(['edit-articles', 'edit-news']));
-    }
-
-    /** @test */
-    public function it_can_determine_that_the_user_has_any_of_the_permissions_via_role()
-    {
-        $this->testUserRole->givePermissionTo('edit-articles');
-
-        $this->testUser->assignRole('testRole');
-
-        $this->assertTrue($this->testUser->hasAnyPermission('edit-news', 'edit-articles'));
-    }
-
-    /** @test */
-    public function it_can_determine_that_user_has_direct_permission()
-    {
-        $this->testUser->givePermissionTo('edit-articles');
-        $this->testUser->assignRole('testRole');
-        $this->testUserRole->givePermissionTo('edit-news');
-        $this->refreshTestUser();
-        $this->assertFalse($this->testUser->hasDirectPermission('edit-news'));
-        $this->assertTrue($this->testUser->hasDirectPermission('edit-articles'));
-        $this->assertEquals(
-            collect(['edit-articles']),
-            $this->testUser->getDirectPermissions()->pluck('name')
-        );
-
-        $this->testUser->revokePermissionTo('edit-articles');
-        $this->refreshTestUser();
-        $this->assertFalse($this->testUser->hasDirectPermission('edit-articles'));
-    }
-
-    /** @test */
-    public function it_can_list_all_the_permissions_via_his_roles()
-    {
-        $roleModel = app(Role::class);
-        $roleModel->findByName('testRole2')->givePermissionTo('edit-news');
-
-        $this->testUserRole->givePermissionTo('edit-articles');
-        $this->testUser->assignRole('testRole', 'testRole2');
-
-        $this->testUser->givePermissionTo('edit-categories');
-
-        $this->assertEquals(
-            collect(['edit-articles', 'edit-news']),
-            $this->testUser->getPermissionsViaRoles()->pluck('name')
-        );
-    }
-
-    /** @test */
-    public function it_can_list_all_the_coupled_permissions_both_directly_and_via_roles()
-    {
-        $this->testUser->givePermissionTo('edit-news');
-
-        $this->testUserRole->givePermissionTo('edit-articles');
-        $this->testUser->assignRole('testRole');
-
-        $this->assertEquals(
-            collect(['edit-articles', 'edit-news']),
-            $this->testUser->getAllPermissions()->pluck('name')
-        );
-    }
-
-    /** @test */
     public function it_can_retrieve_role_names()
     {
         $this->testUser->assignRole('testRole', 'testRole2');
@@ -476,17 +323,6 @@ class HasRolesTest extends TestCase
         $this->assertEquals(
             collect(['testRole', 'testRole2']),
             $this->testUser->getRoleNames()
-        );
-    }
-
-    /** @test */
-    public function it_can_retrieve_permission_names()
-    {
-        $this->testUser->givePermissionTo('edit-articles', 'edit-news');
-
-        $this->assertEquals(
-            collect(['edit-articles', 'edit-news']),
-            $this->testUser->getPermissionNames()
         );
     }
 
@@ -511,31 +347,5 @@ class HasRolesTest extends TestCase
         $this->assertFalse($this->testUser->hasRole('testRole'));
 
         $this->assertFalse($this->testUser->hasRole('testRole2'));
-    }
-
-    /** @test */
-    public function it_can_assign_and_remove_a_role_on_a_permission()
-    {
-        $this->testUserPermission->assignRole('testRole');
-
-        $this->assertTrue($this->testUserPermission->hasRole('testRole'));
-
-        $this->testUserPermission->removeRole('testRole');
-
-        $this->refreshTestUserPermission();
-
-        $this->assertFalse($this->testUserPermission->hasRole('testRole'));
-    }
-
-    /** @test */
-    public function it_can_sync_roles_from_a_string_on_a_permission()
-    {
-        $this->testUserPermission->assignRole('testRole');
-
-        $this->testUserPermission->syncRoles('testRole2');
-
-        $this->assertFalse($this->testUserPermission->hasRole('testRole'));
-
-        $this->assertTrue($this->testUserPermission->hasRole('testRole2'));
     }
 }
